+++ test/compat/regexp-engine/runner.js
@@ -21,7 +21,7 @@ import * as unix from "./unix.test.js";
 import * as win from "./win.test.js";
 
 try {
-  // eslint-disable-next-line prefer-regex-literals, no-invalid-regexp, no-new, regexp/no-non-standard-flag, regexp/require-unicode-regexp
+  // eslint-disable-next-line prefer-regex-literals, no-invalid-regexp, no-new, regexp/no-non-standard-flag
   new RegExp("", "l");
 } catch {
   throw new Error(