+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 ## [Unreleased]
 
 - Correct identifying inherited `shell` before Node.js v22.0.0. ([#2447])
+- Expand support for `--enable-experimental-regexp-engine`. ([#2436])
 
 ## [2.1.10] - 2026-03-10
 
@@ -396,6 +397,7 @@ Versioning].
 [#2387]: https://github.com/ericcornelissen/shescape/pull/2387
 [#2388]: https://github.com/ericcornelissen/shescape/pull/2388
 [#2410]: https://github.com/ericcornelissen/shescape/pull/2410
+[#2436]: https://github.com/ericcornelissen/shescape/pull/2436
 [#2447]: https://github.com/ericcornelissen/shescape/pull/2447
 [552e8ea]: https://github.com/ericcornelissen/shescape/commit/552e8eab56861720b1d4e5474fb65741643358f9
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/